requirements.txt
Django==1.10.3
django-debug-toolbar==1.6
django-bootstrap3==7.1.0
django-bower==5.2.0
django-compressor==2.1
Pillow==3.4.1
psycopg2==2.6.2
django-resized==0.3.5
dj-database-url==0.4.2
gunicorn==19.6.0
whitenoise==3.2
mysqlclient==1.3.10
python-decouple==3.0
django-test-plus==1.0.17